requirements.txt
mistune==3.2.0
python-dotenv==1.0.1
Pillow>=10.0.0
PyYAML>=6.0.1
beautifulsoup4>=4.12.0
websocket-client>=1.8.0
playwright>=1.52.0
imageio-ffmpeg>=0.6.0
